add '--preserve PATTERN', to prevent removal of files/directories a previously released change to --clean-destination caused lazygal to start removing unknown directories, as well as files, in the destination directory.  thus there is no longer a way to protect files which should be kept.
this new option lets a user specify pathnames (via file match pattern)
which should be preserved.  such a mechanism can be useful, for
instance, if the HTML in an album description or in an image caption
wishes to refer to an icon or other document which is not part of the
gallery sources.

unlike most commandline options, the "--preserve" arguments append their
values to the 'preserve' configuration value, rather than replacing
it.  the implementation currently requires a dummy "preserve_args"
config value in defaults.conf.